capsules: system: move capsules to capsules-system
These trait implementations do not need to be in the core kernel.
