+++ capsules/system/src/lib.rs
@@ -1,0 +1,10 @@
+// Licensed under the Apache License, Version 2.0 or the MIT License.
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+// Copyright Tock Contributors 2024.
+
+#![forbid(unsafe_code)]
+#![no_std]
+
+pub mod process_checker;
+pub mod process_policies;
+pub mod process_printer;
